# server/api/api.yaml
openapi: "3.0.0"
info:
  version: 0.6.0
  title: Playbooks API
  contact:
    name: Mattermost
    url: https://mattermost.com/
    email: support@mattermost.com
servers:
  - url: http://localhost:8065/plugins/playbooks/api/v0
paths:
...
  /playbooks:
    get:
      summary: List all playbooks
      description: Retrieve a paged list of playbooks, filtered by team, and sorted by title, number of stages or number of steps.
      operationId: getPlaybooks
      security:
        - BearerAuth: []
      tags:
        - Playbooks
      parameters:
        - name: team_id
          in: query
          description: ID of the team to filter by.
          required: true
          example: 08fmfasq5wit3qyfmq4mjk0rto
          schema:
            type: string
        - name: page
          in: query
          description: Zero-based index of the page to request.
          required: false
          example: 3
          schema:
            type: integer
            format: int32
            default: 0
        - name: per_page
          in: query
          description: Number of playbooks to return per page.
          required: false
          example: 50
          schema:
            type: integer
            format: int32
            default: 1000
        - name: sort
          in: query
          description: Field to sort the returned playbooks by title, number of stages or total number of steps.
          required: false
          example: stages
          schema:
            type: string
            default: title
            enum:
              - title
              - stages
              - steps
        - name: direction
          in: query
          description: Direction (ascending or descending) followed by the sorting of the playbooks.
          required: false
          example: asc
          schema:
            type: string
            default: asc
            enum:
              - desc
              - asc
      x-codeSamples:
        - lang: curl
          source: |
            curl -X GET 'http://localhost:8065/plugins/playbooks/api/v0/playbooks?team_id=08fmfasq5wit3qyfmq4mjk0rto&sort=title&direction=asc' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'
      responses:
        200:
          description: A paged list of playbooks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybookList"
        400:
          $ref: "#/components/schemas/400"
        403:
          $ref: "#/components/schemas/403"
        500:
          $ref: "#/components/schemas/500"
    post:
      summary: Create a playbook
      operationId: createPlaybook
      security:
        - BearerAuth: []
      tags:
        - Playbooks
      requestBody:
        description: Playbook
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - team_id
                - create_public_playbook_run
                - checklists
                - member_ids
              properties:
                title:
                  type: string
                  description: The title of the playbook.
                  example: Cloud PlaybookRuns
                description:
                  type: string
                  description: The description of the playbook.
                  example: A playbook to follow when there is a playbook run regarding the availability of the cloud service.
                team_id:
                  type: string
                  description: The identifier of the team where the playbook is in.
                  example: p03rbi6viyztysbqnkvcqyel2i
                create_public_playbook_run:
                  type: boolean
                  description: A boolean indicating whether the playbook runs created from this playbook should be public or private.
                  example: true
                checklists:
                  type: array
                  description: The stages defined by this playbook.
                  items:
                    type: object
                    required:
                      - title
                      - items
                    properties:
                      title:
                        type: string
                        description: The title of the checklist.
                        example: Triage issue
                      items:
                        type: array
                        description: The list of tasks to do.
                        items:
                          type: object
                          required:
                            - title
                          properties:
                            title:
                              type: string
                              description: The title of the checklist item.
                              example: Gather information from customer.
                            command:
                              type: string
                              description: The slash command associated with this item. If the item has no slash command associated, this is an empty string
                              example: /opsgenie on-call
                            description:
                              type: string
                              description: A detailed description of the checklist item, formatted with Markdown.
                              example: Ask the customer for more information in [Zendesk](https://www.zendesk.com/).
                member_ids:
                  description: The identifiers of all the users that are members of this playbook.
                  type: array
                  items:
                    type: string
                    description: User ID of the playbook member.
                    example: ilh6s1j4yefbdhxhtlzt179i6m
                broadcast_channel_ids:
                  description: The IDs of the channels where all the status updates will be broadcasted. The team of the broadcast channel must be the same as the playbook's team.
                  type: array
                  items:
                    type: string
                    description: ID of the broadcast channel.
                    example: 2zh7rpashwfwapwaqyslmhwbax
                invited_user_ids:
                  description: A list with the IDs of the members to be automatically invited to the playbook run's channel as soon as the playbook run is created.
                  type: array
                  items:
                    type: string
                    description: User ID of the member to be invited.
                    example: 01kidjn9iozv7bist427w4gkjo
                invite_users_enabled:
                  description: Boolean that indicates whether the members declared in invited_user_ids will be automatically invited.
                  type: boolean
                  example: true
                default_owner_id:
                  description: User ID of the member that will be automatically assigned as owner as soon as the playbook run is created. If the member is not part of the playbook run's channel or is not included in the invited_user_ids list, they will be automatically invited to the channel.
                  type: string
                  example: 9dtruav6d9ce3oqnc5pwhtqtfq
                default_owner_enabled:
                  description: Boolean that indicates whether the member declared in default_owner_id will be automatically assigned as owner.
                  type: string
                  example: true
                announcement_channel_id:
                  description: ID of the channel where the playbook run will be automatically announced as soon as the playbook run is created.
                  type: string
                  example: 8iofau5swv32l6qtk3vlxgobta
                announcement_channel_enabled:
                  description: Boolean that indicates whether the playbook run creation will be announced in the channel declared in announcement_channel_id.
                  type: boolean
                  example: true
                webhook_on_creation_url:
                  description: An absolute URL where a POST request will be sent as soon as the playbook run is created. The allowed protocols are HTTP and HTTPS.
                  type: string
                  example: https://httpbin.org/post
                webhook_on_creation_enabled:
                  description: Boolean that indicates whether the webhook declared in webhook_on_creation_url will be automatically sent.
                  type: boolean
                  example: true
                webhook_on_status_update_url:
                  description: An absolute URL where a POST request will be sent as soon as the playbook run's status is updated. The allowed protocols are HTTP and HTTPS.
                  type: string
                  example: https://httpbin.org/post
                webhook_on_status_update_enabled:
                  description: Boolean that indicates whether the webhook declared in webhook_on_status_update_url will be automatically sent.
                  type: boolean
                  example: true
      x-codeSamples:
        - lang: curl
          source: |
            curl -X POST 'http://localhost:8065/plugins/playbooks/api/v0/playbooks' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'\
            -d '{"title": "Cloud PlaybookRuns", "description": "A playbook to follow when there is a playbook run regarding the availability of the cloud service.", "team_id": "p03rbi6viyztysbqnkvcqyel2i","create_public_playbook_run": true,"checklists": [{"title": "Triage issue","items": [{"title": "Gather information from customer."}]}]}'
      callbacks:
        playbookRunCreation:
          "{$request.body#/webhook_on_creation_url}":
            post:
              summary: PlaybookRun's creation outgoing webhook.
              description: When a playbook run is created with this playbook, a POST request is sent to the URL configured in webhook_on_creation_url. The webhook is considered successful if your server returns a response code within the 200-299 range. Otherwise, the webhook is considered failed, and a warning message is posted in the playbook run's channel. No retries are made.
              operationId: webhookOncreation
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/WebhookOnCreationPayload"
              responses:
                "2XX":
                  description: Your server returns a 2XX code if it successfully received the request.
        playbookRunStatusUpdate:
          "{$request.body#/webhook_on_status_update_url}":
            post:
              summary: PlaybookRun's status update outgoing webhook.
              description: When a playbook run's status is updated, a POST request is sent to the URL configured in webhook_on_status_update_url. The webhook is considered successful if your server returns a response code within the 200-299 range. Otherwise, the webhook is considered failed, and a warning message is posted in the playbook run's channel. No retries are made.
              operationId: webhookOnStatusUpdate
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      $ref: "#/components/schemas/WebhookOnStatusUpdatePayload"
              responses:
                "2XX":
                  description: Your server returns a 2XX code if it successfully received the request.
      responses:
        201:
          description: ID of the created playbook.
          headers:
            Location:
              description: Location of the created playbook.
              schema:
                type: string
                example: /api/v0/playbook/cdl5o0tjcp5rqlpjidhobj64nd
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: iz0g457ikesz55dhxcfa0fk9yy
                required:
                  - id
        400:
          $ref: "#/components/schemas/400"
        403:
          $ref: "#/components/schemas/403"
        500:
          $ref: "#/components/schemas/500"
  /playbooks/{id}:
    get:
      summary: Get a playbook
      operationId: getPlaybook
      security:
        - BearerAuth: []
      tags:
        - Playbooks
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the playbook to retrieve.
          schema:
            type: string
            example: iz0g457ikesz55dhxcfa0fk9yy
      x-codeSamples:
        - lang: curl
          source: |
            curl -X GET 'http://localhost:8065/plugins/playbooks/api/v0/playbooks/iz0g457ikesz55dhxcfa0fk9yy' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'
      responses:
        200:
          description: Playbook.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playbook"
        403:
          $ref: "#/components/schemas/403"
        500:
          $ref: "#/components/schemas/500"
    put:
      summary: Update a playbook
      operationId: updatePlaybook
      security:
        - BearerAuth: []
      tags:
        - Playbooks
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the playbook to update.
          schema:
            type: string
            example: iz0g457ikesz55dhxcfa0fk9yy
      requestBody:
        description: Playbook payload
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Playbook"
      x-codeSamples:
        - lang: curl
          source: |
            curl -X PUT 'http://localhost:8065/plugins/playbooks/api/v0/playbooks/iz0g457ikesz55dhxcfa0fk9yy' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'\
            -d '{"title": "Playbook","team_id": "ni8duypfe7bamprxqeffd563gy","create_public_playbook_run": true,"checklists": [{"title": "Title","items": [{"title": "Title"}]}]}'
      responses:
        200:
          description: Playbook succesfully updated.
        400:
          $ref: "#/components/schemas/400"
        403:
          $ref: "#/components/schemas/403"
        500:
          $ref: "#/components/schemas/500"
    delete:
      summary: Delete a playbook
      operationId: deletePlaybook
      security:
        - BearerAuth: []
      tags:
        - Playbooks
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the playbook to delete.
          schema:
            type: string
            example: iz0g457ikesz55dhxcfa0fk9yy
      x-codeSamples:
        - lang: curl
          source: |
            curl -X DELETE 'http://localhost:8065/plugins/playbooks/api/v0/playbooks/iz0g457ikesz55dhxcfa0fk9yy' \
            -H 'Authorization: Bearer 9g64ig7q9pds8yjz8rsgd6e36e'
      responses:
        204:
          description: Playbook successfully deleted.
        403:
          $ref: "#/components/schemas/403"
        500:
          $ref: "#/components/schemas/500"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    400:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: The request is malformed.
    403:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: Access to the resource is forbidden for this user.
...
    500:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: There was an internal error in the server.
...
    PlaybookList:
      type: object
      properties:
        total_count:
          type: integer
          description: The total number of playbooks in the list, regardless of the paging.
          format: int32
          example: 305
        page_count:
          type: integer
          description: The total number of pages. This depends on the total number of playbooks in the database and the per_page parameter sent with the request.
          format: int32
          example: 2
        has_more:
          type: boolean
          description: A boolean describing whether there are more pages after the currently returned.
          example: true
        items:
          type: array
          description: The playbooks in this page.
          items:
            $ref: "#/components/schemas/Playbook"
...
    Error:
      type: object
      required:
        - error
        - details
      properties:
        error:
          type: string
          description: A message with the error description.
          example: Error retrieving the resource.
        details:
          type: string
          description: Further details on where and why this error happened.
          example: Specific details about the error, depending on the case.
    WebhookOnCreationPayload:
      allOf:
        - $ref: "#/components/schemas/PlaybookRun"
        - type: object
          properties:
            channel_url:
              type: string
              description: Absolute URL to the playbook run's channel.
              example: http://example.com/ad-1/channels/channel-name
            details_url:
              type: string
              description: Absolute URL to the playbook run's details.
              example: http://example.com/ad-1/playbooks/runs/playbookRunID
    WebhookOnStatusUpdatePayload:
      allOf:
        - $ref: '#/components/schemas/PlaybookRun'
        - type: object
          properties:
            channel_url:
              type: string
              description: Absolute URL to the playbook run's channel.
              example: http://example.com/ad-1/channels/channel-name
            details_url:
              type: string
              description: Absolute URL to the playbook run's details.
              example: http://example.com/ad-1/playbooks/runs/playbookRunID